+++ rdbbeat/exceptions.py
